tried different noise models
I tried to introduce noise in just some stretches of the polymer,
mimicking conformation preservation of some domains of the chromosome.
Expected that after the reconstruction the precision  for those
conserved domains will be higher than for the other, but that did not
work, the error is evenly spread along the chain.